+++ BusinessLayer/DTOs/Product/UpdateProductDto.cs
@@ -10,7 +10,7 @@ public class UpdateProductDto
     public decimal UnitPrice { get; set; }
     public int QuantityInStock { get; set; }
     public CategoryEnum? Category { get; set; }
-    public int SupId { get; set; }
+    public long SupId { get; set; }
 }
 public static class UpdateMappingExtensions
 {
